+++ lib/routes.js
@@ -4,7 +4,7 @@ const path = require('path')
 const yaml = require('yaml')
 const fastifyStatic = require('@fastify/static')
 const rfdc = require('rfdc')()
-const swaggerInitializer = require('./swaggerInitializer')
+const swaggerInitializer = require('./swagger-initializer')
 
 // URI prefix to separate static assets for swagger UI
 const staticPrefix = '/static'
