+++ lib/swagger-initializer.js
